+++ lib/src/models/create_config.dart
@@ -1,4 +1,5 @@
 import 'package:arise/src/models/architecture.dart';
+import 'package:arise/src/models/networking.dart';
 import 'package:arise/src/models/routing.dart';
 import 'package:arise/src/models/state_management.dart';
 
@@ -8,10 +9,12 @@ class CreateConfig {
     required this.architecture,
     required this.stateManagement,
     required this.routing,
+    required this.networking,
   });
 
   final String projectName;
-  final StateManagement stateManagement;
   final Architecture architecture;
+  final StateManagement stateManagement;
   final Routing routing;
+  final Networking networking;
 }