+++ lib/src/models/networking.dart
@@ -1,0 +1,9 @@
+enum Networking {
+  none('None', 'none'),
+  dio('Dio', 'dio');
+
+  const Networking(this.label, this.templateName);
+
+  final String label;
+  final String templateName;
+}